+++ src/main/java/com/example/Consultorio/repositories/PacienteRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PacienteRepository extends CrudRepository<PacienteModel,Long> {
+public interface PacienteRepository extends CrudRepository<PacienteModel, Integer> {
 }
